20 URL columns -> one URLs column
The spreadsheet had 20 URL columns.  I altered the spreadsheet to have
one URLs column with all of the URLs in it, separated by *double* spaces
(because some URLs contain a single space, which they really shouldn't),
and I altered idno.xsl to tokenize the URLs column to have the same
functionality as the 20 URL columns.  I also cleaned the spreadsheet of
double-spaces in name columns, and the CBSC en-Full column now contains
a semi-colon separated list of CBSC names where there are multiple.
This breaks the CBSC en-Link column formula, but I haven't fixed that
yet.