only one source's info for events
It was decided that the easiest way to deal with contradictory date
information for events was to include only the information from the most
reliable source.  This XSLT now prefers the first column which has
information, and it depends upon the columns being ordered from most
reliable to least reliable.