places in names, fixed bug in comma titles
Now places which occur in "Barsoum en Name Place" column are wrapped in
a <placeName> element with a @ref pointing to the URI in the "Barsoum en
Name Place URI" column.  Also, I found a bug in the name-parts template
when it handled comma-separated titles, namely that it discarded all of
the markup generated by the name-parts-comma-separated template by
virtue of using <xsl:value-of> rather than <xsl:copy-of>.  Swapping it
to <xsl:copy-of> generated ~30 cases where a column title was output
into the TEI (which does not validate), which was caused by there being
a name in the spreadsheet which was not divided into given/family/title
names.  I still do not know what caused this brittle behavior, and it
remains.  By dividing the remaining names, column names are no longer
output into the TEI, for whatever reason.