extra space in Arabic names & person/563
A bunch of TEI files had an extra space following an Arabic name due to
a spreadsheet data entry error.  This was fixed in the spreadsheet, and
the spaces were removed.  This resulted in correct tagging of a title in
person/37.  A family name for person/563 was also not tagged as a family
name due to a spreadsheet data entry error; this has been fixed.