closed logical gap in names.xsl line 116
Though I don't have any example of the XSL misbehaving, in theory it was
possible, if the Authorized_syr-Source column pointed to Abdisho or
Barsoum and the Authorized_syr name columns contained data which didn't
match the designated authorized name to end up with two Syriac names
designated as headwords.  I fixed this.