fix GEDSH-Style supersedes GEDSH code
The earlier code was supposed to only make a GEDSH name the
syriaca-headword if there was no GEDSH-style name to supersede it, but I
wrote that code incorrectly, and it did no such thing.  I fixed that
bug, so now GEDSH names can only be labeled syriaca-headword if there is
no GS_en-Full names to supersede it.