Added list of 40 possible malformed names
Notes:
This list provides a concatenated list of any strings (other than a
single comma) that are in the text node of a persName element but not
in a name part element.
Note: Some files have more than one malformed persName text node and
thus are listed more than once for each persName element with the xmlid
for the persName noted.

A separate list of names with just commas could be generated if we
thing these are needed.