removed "syriaca-anglicized" from duplicate names
Sometimes the name in the "Other en-" column was also contained in
another column, sometimes even the GEDSH or GS columns.  This was
replacing "syriaca-headword" with "syriaca-anglicized," and was marking
names from CBSC or Barsoum as "syriaca-anglicized."  I removed
"syriaca-anglicized" from these names and only applied it to names which
only exist in the "Other en" columns.